Rename utils to common_utils
