+++ src/handle_get_printable_amount.c
@@ -4,7 +4,7 @@
 #include "swap_utils.h"
 #include "handle_get_printable_amount.h"
 #include "shared_context.h"
-#include "utils.h"
+#include "common_utils.h"
 #include "uint256.h"
 #include "string.h"
