+++ src/handle_swap_sign_transaction.c
@@ -4,7 +4,7 @@
 #include "swap_utils.h"
 #include "handle_swap_sign_transaction.h"
 #include "shared_context.h"
-#include "utils.h"
+#include "common_utils.h"
 #ifdef HAVE_NBGL
 #include "nbgl_use_case.h"
 #endif  // HAVE_NBGL
