+++ src_bagl/ui_flow.c
@@ -1,7 +1,7 @@
 #include "shared_context.h"
 #include "ui_callbacks.h"
 #include "common_ui.h"
-#include "utils.h"
+#include "common_utils.h"
 
 #define ENABLED_STR   "Enabled"
 #define DISABLED_STR  "Disabled"
