+++ src_bagl/ui_flow_signTx.c
@@ -1,7 +1,7 @@
 #include "shared_context.h"
 #include "ui_callbacks.h"
 #include "chainConfig.h"
-#include "utils.h"
+#include "common_utils.h"
 #include "feature_signTx.h"
 #include "network.h"
 #include "eth_plugin_handler.h"
