+++ src_features/provideNFTInformation/cmd_provideNFTInfo.c
@@ -3,7 +3,7 @@
 #include "shared_context.h"
 #include "apdu_constants.h"
 #include "tokens.h"
-#include "utils.h"
+#include "common_utils.h"
 #include "common_ui.h"
 #include "os_io_seproxyhal.h"
 #include "network.h"
