+++ src_features/setPlugin/cmd_setPlugin.c
@@ -4,7 +4,7 @@
 #include "eth_plugin_interface.h"
 #include "eth_plugin_internal.h"
 #include "plugin_utils.h"
-#include "utils.h"
+#include "common_utils.h"
 #include "common_ui.h"
 #include "os_io_seproxyhal.h"
 #include "network.h"
