+++ src_features/signMessageEIP712_v0/cmd_signMessage712.c
@@ -1,6 +1,6 @@
 #include "shared_context.h"
 #include "apdu_constants.h"
-#include "utils.h"
+#include "common_utils.h"
 #include "common_ui.h"
 #include "common_712.h"
