+++ src_plugins/erc1155/erc1155_ui.c
@@ -4,7 +4,7 @@
 #include "erc1155_plugin.h"
 #include "eth_plugin_internal.h"
 #include "eth_plugin_interface.h"
-#include "utils.h"
+#include "common_utils.h"
 
 static void set_approval_for_all_ui(ethQueryContractUI_t *msg, erc1155_context_t *context) {
     switch (msg->screenIndex) {
