+++ src_plugins/erc20/erc20_plugin.c
@@ -4,7 +4,7 @@
 #include "shared_context.h"
 #include "plugin_utils.h"
 #include "ethUstream.h"
-#include "utils.h"
+#include "common_utils.h"
 
 typedef enum { ERC20_TRANSFER = 0, ERC20_APPROVE } erc20Selector_t;
