+++ src_plugins/eth2/eth2_plugin.c
@@ -4,7 +4,7 @@
 #include "eth_plugin_internal.h"
 #include "eth_plugin_handler.h"
 #include "shared_context.h"
-#include "utils.h"
+#include "common_utils.h"
 
 void getEth2PublicKey(uint32_t *bip32Path, uint8_t bip32PathLength, uint8_t *out);
