+++ src_plugins/starkware/starkware_plugin.c
@@ -5,7 +5,7 @@
 #include "eth_plugin_internal.h"  // TODO : rewrite as independant code
 #include "stark_utils.h"
 #include "plugin_utils.h"
-#include "utils.h"
+#include "common_utils.h"
 #include "apdu_constants.h"
 
 #ifdef HAVE_STARKWARE
